# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Search Platform API",
        "description": "API Documentation for Job Search Platform Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users",
                "description": "Returns a paginated list of users with optional filtering.",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filter by role (JobSeeker, Employer, Admin)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (active, blocked, unverified)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new user",
                "description": "Create a new user (by admin).",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "JobSeeker",
                                            "Employer",
                                            "Admin"
                                        ]
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/applications/{application}/ai-match": {
            "get": {
                "tags": [
                    "AI Analytics"
                ],
                "summary": "Get AI match analysis for an application",
                "description": "Returns the AI screening match score, notes, and detailed CV-Job compatibility breakdown.",
                "operationId": "getAiMatch",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "Application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AI match details"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Application not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cvs/{cv}/analysis": {
            "get": {
                "tags": [
                    "AI Analytics"
                ],
                "summary": "Get AI analysis of a CV",
                "description": "Returns scores, strengths, gaps, and improvement recommendations for a CV.",
                "operationId": "getCvAnalysis",
                "parameters": [
                    {
                        "name": "cv",
                        "in": "path",
                        "description": "CV ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CV analysis results"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "No analysis found for this CV"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cvs/{cv}/skill-gaps": {
            "get": {
                "tags": [
                    "AI Analytics"
                ],
                "summary": "Get skill gap analysis for a CV",
                "description": "Returns a list of skills the job seeker is missing compared to market demands.",
                "operationId": "getCvSkillGaps",
                "parameters": [
                    {
                        "name": "cv",
                        "in": "path",
                        "description": "CV ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "job_ad_id",
                        "in": "query",
                        "description": "Filter by specific job ad",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill gap analysis results"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/jobs/{jobId}/match-score": {
            "get": {
                "tags": [
                    "AI Analytics"
                ],
                "summary": "Get match score before applying",
                "description": "Returns the CV match score for a specific job without actually applying.",
                "operationId": "getPreApplyMatchScore",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job Ad ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cv_id",
                        "in": "query",
                        "description": "CV ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match score details"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Job or CV not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get my applications",
                "description": "Returns a list of applications submitted by the current job seeker.",
                "operationId": "getMyApplications",
                "responses": {
                    "200": {
                        "description": "List of applications"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Submit a job application",
                "description": "Apply to a job.",
                "operationId": "submitApplication",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "job_id",
                                    "cv_id"
                                ],
                                "properties": {
                                    "job_id": {
                                        "type": "integer"
                                    },
                                    "cv_id": {
                                        "type": "integer"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications/{id}": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get application details",
                "description": "Get a specific application.",
                "operationId": "getApplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications/auto-apply": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Auto apply to a matching job",
                "description": "Finds the best matching active job for the provided CV and applies automatically.",
                "operationId": "autoApply",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cv_id"
                                ],
                                "properties": {
                                    "cv_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Auto application result"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications/{id}/withdraw": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Withdraw an application",
                "description": "Withdraw an application.",
                "operationId": "withdrawApplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application withdrawn"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/employer/applications/{id}/status": {
            "put": {
                "tags": [
                    "Employer",
                    "Applications"
                ],
                "summary": "Update application status",
                "description": "Update application status (for employers).",
                "operationId": "updateApplicationStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Pending",
                                            "Reviewed",
                                            "Shortlisted",
                                            "Interviewing",
                                            "Offered",
                                            "Hired",
                                            "Rejected"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/login/{provider}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Social Login Redirect",
                "description": "Returns the redirection URL for social login (Google/LinkedIn).",
                "operationId": "socialLoginRedirect",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "linkedin"
                            ]
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Optional role to assign if a new account is created (JobSeeker, Employer).",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "JobSeeker",
                                "Employer"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect URL generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "example": "https://accounts.google.com/..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Logout current user.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}